Pass Arrow Schema Pointer From Python To C
