+++ apis/python/pyproject.toml
@@ -1,5 +1,6 @@
 [build-system]
 requires = [
+    "pyarrow",
     "pybind11[global]>=2.10.0",
     "setuptools>=65.5.1",
     "wheel>=0.37.1",
